replaced uint8_t to size_t as of now, as it is creating problems while reading from sstream

+++ GD_Assignment2/headers/Config.hpp
@@ -14,21 +14,21 @@ public:
 
 	std::string fontFilePath;
 	size_t fontSize;
-	uint8_t fontR, fontG, fontB;
+	size_t fontR, fontG, fontB;
 
 	size_t playerSR; // Shape Radius
 	size_t playerCR; // Collision Radius
 	float playerSpeed;
-	uint8_t playerFR, playerFG, playerFB; // Fill Color
-	uint8_t playerOR, playerOG, playerOB; // Outline Color
+	size_t playerFR, playerFG, playerFB; // Fill Color
+	size_t playerOR, playerOG, playerOB; // Outline Color
 	size_t playerOT; // Outline Thickness
 	size_t playerVertices; // Shape vertices count or point count
 	
 	size_t enemySR; // Shape Radius
 	size_t enemyCR; // Collision Radius
 	float enemySMin, enemySMax; // Speed range
-	uint8_t enemyFR, enemyFG, enemyFB; // Fill Color
-	uint8_t enemyOR, enemyOG, enemyOB; // Outline Color
+	size_t enemyFR, enemyFG, enemyFB; // Fill Color
+	size_t enemyOR, enemyOG, enemyOB; // Outline Color
 	size_t enemyOT; // Outline Thickness
 	size_t enemyVertMin, enemyVertMax; // Shape vertices count or point count
 	size_t enemyL; // Small lifespan
@@ -37,8 +37,8 @@ public:
 	size_t bulletSR; // Shape Radius
 	size_t bulletCR; // Collision Radius
 	float bulletSpeed;
-	uint8_t bulletFR, bulletFG, bulletFB; // Fill Color
-	uint8_t bulletOR, bulletOG, bulletOB; // Outline Color
+	size_t bulletFR, bulletFG, bulletFB; // Fill Color
+	size_t bulletOR, bulletOG, bulletOB; // Outline Color
 	size_t bulletOT; // Outline Thickness
 	size_t bulletVertices; // Shape vertices count or point count
 	size_t bulletLifespan;
